Add: db attribute in parent class Page
- Moved method connect to db from Home child class to Page parent class.
- Updated return_table_as_df method in user_db module.
- Removed unused read_table method in user_db module.
